fix(FCB-19): fix collisions ids profile and request
