+++ src/widgets/modals/components/ImportModal/components/ImportModalBody/constants.ts
@@ -1,10 +1,8 @@
-export const TOOLTIP_TITLE = `Typing of JSON field values: 
+export const TOOLTIP_TITLE = `Typing of JSON field values:
 [
   {
-    "id": string,
     "requestHeaders": [
       {
-        "id": number,
         "name": string,
         "value": string,
         "disabled": boolean
